FIX: tensor dtype

+++ srcs/model.py
@@ -42,10 +42,10 @@ class QTrainer:
         game_over: bool | tuple[bool]
     ):
 
-        state = torch.tensor(state)
-        action = torch.tensor(action)
-        reward = torch.tensor(reward)
-        next_state = torch.tensor(next_state)
+        state = torch.tensor(state, dtype=torch.float32)
+        action = torch.tensor(action, dtype=torch.int32)
+        reward = torch.tensor(reward, dtype=torch.float32)
+        next_state = torch.tensor(next_state, dtype=torch.float32)
 
         if len(state.shape) == 1:
             state = torch.unsqueeze(state, 0)
